requirements.txt
langserve[all]
langchain-groq
langchain[pdf]
pypdf
python-dotenv
uvicorn
gunicorn
langgraph
python-multipart
requests
pytest
grandalf
fastapi
langchain
groq
beautifulsoup4
pydantic-settings
playwright
PyMuPDF